+++ test/format_test.dart
@@ -73,12 +73,6 @@ void main() {
       });
     });
 
-    test('toString()', () {
-      final iCal = ICalendar.fromLines(valid);
-      final str = iCal.toString();
-      expect(str.contains('iCalendar - VERSION: 2.0 - PRODID: '), true);
-    });
-
     group('jsonEncodable()', () {
       test('IcsDateTime', () {
         final icsDt = IcsDateTime(dt: DateTime.now().toIso8601String());
@@ -205,9 +199,9 @@ void main() {
       final exdate =
           iCal.data.firstWhere((e) => e.containsKey('exdate'))['exdate'];
       expect(exdate, [
-        IcsDateTime(dt: "20220415T145500", tzid: "Europe/Zurich"),
-        IcsDateTime(dt: "20220527T145500", tzid: "Europe/Zurich"),
-        IcsDateTime(dt: "20220422T145500", tzid: "Europe/Zurich"),
+        const IcsDateTime(dt: "20220415T145500", tzid: "Europe/Zurich"),
+        const IcsDateTime(dt: "20220527T145500", tzid: "Europe/Zurich"),
+        const IcsDateTime(dt: "20220422T145500", tzid: "Europe/Zurich"),
       ]);
     });
   });